size_t

+++ gulkov.mikhail/I4/CountPositiveCols.cpp
@@ -2,7 +2,7 @@
 
 size_t countPositiveCols(const int *array, size_t cols, size_t rows)
 {
-  unsigned int positive_cols = 0;
+  size_t positive_cols = 0;
   for (size_t i = 0; i < cols; i++)
   {
     for (size_t j = 0; j < rows; j++)
